requirements.txt
Django==1.6.5
bpython==0.13
django-braces==1.4.0
djangorestframework==2.3.12
boto==2.32.0
